refactored and cleaned inventory states and hud

+++ Link/LinkSprites/LinkFacingSprites.cs
@@ -36,7 +36,7 @@ namespace Sprites
 
             // Create source and destination rectangles
             sourceRectangle = new Rectangle(71, 11, 12, 16); // Store the current location on the spritesheet to get a sprite from
-            destinationRectangle = new Rectangle(xPosition, yPosition, sourceRectangle.Width * 2, sourceRectangle.Height * 2); // Where to draw on screen
+            destinationRectangle = new Rectangle((int)xPosition, (int)yPosition, sourceRectangle.Width * 2, sourceRectangle.Height * 2); // Where to draw on screen
         }
         public void Update()
         {
